# [decoder] use typed models for msgpack decoding