--- a/.gitattributes
+++ b/.gitattributes
@@ -1,0 +1,2 @@
+# Binary or generated files
+model/typed_msgpack.go binary