… model/typed_msgpack.go