@palazzem 2017-01-02
OK @LeoCavaille, we can delay this PR until we've tried it for a lot of time with a successful fuzzy, real data testing.