@clutchski 2017-01-03
though it feels like protobuf should handle this case just as well no?