@palazzem 2017-01-05
Trying another PR since we can get a better performance with the same approach.